requirements.txt
uuid
redis
types-redis
apache-airflow
lxml
bs4
bcrypt
python-dotenv
fastapi
uvicorn
sqlalchemy
snowflake-sqlalchemy
pydantic
streamlit
boto3
snowflake
matplotlib
pandas
tabula-py
pypdf2
langchain_openai
langchain